+++ test/bigtest/mocks/index.js
@@ -1,3 +1,4 @@
 export * from './job-executions';
 export * from './jobs-logs';
 export * from './job-log';
+export * from './associated-job-profiles';